mm: Use size_t for malloc
That is what the standard says to do

+++ include/mm/mm.h
@@ -1,10 +1,11 @@
 #ifndef MM_MM_H_INCLUDED
 #define MM_MM_H_INCLUDED
 
+#include <stddef.h>
 #include <stdint.h>
 
 void init_heap(void) __attribute__((section(".kernel")));
-void *kmalloc(uint32_t size) __attribute__((malloc,section(".kernel")));
+void *kmalloc(size_t size) __attribute__((malloc,section(".kernel")));
 void kfree(void *address) __attribute__((section(".kernel")));
 
 uint32_t mm_space(void) __attribute__((section(".kernel")));
