+++ include/stdlib.h
@@ -1,9 +1,10 @@
 #ifndef STDLIB_H_INCLUDED
 #define STDLIB_H_INCLUDED
 
+#include <stddef.h>
 #include <stdint.h>
 
-void *malloc(uint32_t size) __attribute__((malloc,section(".kernel")));
+void *malloc(size_t size) __attribute__((malloc,section(".kernel")));
 void free(void *address) __attribute__((section(".kernel")));
 void abort(void) __attribute__((section(".kernel")));
 char *strndup(char *str, int n);
